Finished the generation algorithm
Images can be generated from source images now. The generated images share a similar color scheme to the source photo (even some of the same shapes), however most of the time it looks like a net of various colors. The algorithm is slow and messy right now, but hopefully that will be remedied in the future.